LayoutTests/imported/w3c/web-platform-tests/html/syntax/speculative-parsing/generated/page-load/meta-referrer-no-referrer-img-src.tentative-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: Error: assert_not_equals: speculative case did not fetch got disallowed value ""


FAIL Speculative parsing, page load: meta-referrer-no-referrer-img-src Unhandled rejection: assert_not_equals: speculative case did not fetch got disallowed value ""
